variable jump constants
after each mapping of a element, the key will be cycled with a variable step each time rather than only 0. the jump constant is equal to the index of the first char of the inSeq in the alphabet

Issue:
this will eventually lead to the cycling breaking into an infinite loop of a certain sequence.