introducing salting and improving cycling
introducing salting which sprinkles in some random elements such that they may cloud how long the orignal message is, as well as confuse any interceptors, and it will make the proccess of breaking the encryption way harder as mistaking one salt value to be a true value will desync the key by cycling, which will disable any interceptors from reading the encrypted message.

issues:
salts are sprinkled in uniformly, which means if a person some how knows exactly how the random values are scattered this may make it easier for them to break the encryption(although it is still borderline impossible to be broken unless the interceptor knows the keyvas the cycling mechenism makes bruteforcing it astronomically long).
I should find a less predictable way to scatter the values rather than uniformly in encremments governed by the value of the variable saltJump, which similar to cyclejump is governed by the key.